pyproject.toml
[tool.poetry]
name = "ttb"
version = "0.1.0"
description = ""
authors = ["DavidGL17 <david.gonzalez@citycable.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.0"
requests = "^2.31.0"
pyyaml = "^6.0"
requests-oauthlib = "^1.3.1"

[tool.poetry.scripts]
ttb = "ttb.main:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
poethepoet = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
lint = "poetry run flake8 ttb tests"
test = "poetry run pytest tests"
format = "poetry run black ttb tests --line-length 120 --check"
run = "poetry run python ttb.ttb:main"